// src/app/models/suppliers.model.ts
export class Suppliers {
  suppliersId: number = 0;
  supplierId: string = '';
  companyName: string = '';
  contactName: string = '';
  emailAddress: string = '';
  address1: string = '';
  address2: string = '';
  city: string = '';
  region: string = '';
  postalCode: string = '';
  country: string = '';
  phone: string = '';
  fax: string = '';
  shippingMethod: string = '';
  leadtimeDays: number= 0;
  paymentTerms: string = '';
  picPath: string = '';
}